Extract seeded RNG helpers into shared seededRandom module
